--- config.json.example
@@ -1,5 +1,0 @@
-{
-  "token": "",
-  "clientId": "",
-  "guildId": ""
-}
