+++ package.json
@@ -1,17 +1,19 @@
 {
   "name": "sovbot",
   "version": "1.2.3",
-  "main": "index.js",
+  "main": "src/index.ts",
   "license": "MIT",
   "scripts": {
     "lint": "eslint .",
     "ready": "node --trace-warnings .",
-    "deploy-commands": "node deploy-commands.js",
+    "deploy-commands": "node src/deploy-commands.ts",
     "start": "yarn run deploy-commands && yarn run ready"
   },
   "devDependencies": {
+    "@tsconfig/node16": "^1.0.3",
     "cz-conventional-changelog": "3.3.0",
-    "eslint": "^8.20.0"
+    "eslint": "^8.20.0",
+    "typescript": "^4.8.4"
   },
   "config": {
     "commitizen": {
