+++ src/config.json.example
@@ -1,0 +1,5 @@
+{
+  "token": "",
+  "clientId": "",
+  "guildId": ""
+}
