+++ src/events/ready.ts
@@ -1,0 +1,7 @@
+module.exports = {
+	name: 'ready',
+	once: true,
+	execute(client) {
+		console.log(`Ready! Logged in as ${client.user.tag}`);
+	},
+};
